backend/main.py
import os
import uuid
import json
from datetime import datetime, timedelta
from typing import List, Optional, Dict, Tuple
from fastapi import FastAPI, UploadFile, File, HTTPException
from pydantic import BaseModel
import numpy as np
import networkx as nx
from dataclasses import asdict

from shared.models import Audit, Certificate, AccessLevel, StabilityProfile, CausalEffects, DriftAlert
from backend.app.stability.audit import run_stability_audit
from backend.app.drift.subscriber import router as drift_router
from backend.app.report.gemini_report import generate_audit_report, AuditReport
from google.cloud import storage

# Local Mechanistic Audit imports
FAIRGUARD_LOCAL_MODE = os.getenv(\"FAIRGUARD_LOCAL_MODE\", \"0\") == \"1\"
if FAIRGUARD_LOCAL_MODE:
    try:
        from backend.app.mechanistic.server import router as mechanistic_router
    except ImportError:
        mechanistic_router = None
else:
    mechanistic_router = None

# Import causal modules
try:
    from backend.app.causal.scm import StructuralCausalModel
    from backend.app.causal.decompose import ctf_de, ctf_ie, ctf_se, total_variation
    from backend.app.causal.gemini_dag import elicit_dag_from_description, build_dag_from_json
except ImportError:
    StructuralCausalModel = None
    ctf_de = ctf_ie = ctf_se = total_variation = None

from fastapi.middleware.cors import CORSMiddleware

app = FastAPI(title="FairGuard Backend", version="1.0.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(drift_router)

if FAIRGUARD_LOCAL_MODE and mechanistic_router:
    app.include_router(mechanistic_router, tags=["Mechanistic"])

# In-memory storage for demo
audits = {}
certificates = {}

# GCS Cache Configuration
BUCKET_NAME = \"fairguard-artifacts\"

def get_cached_report(audit_id: str) -> Optional[dict]:
    try:
        client = storage.Client()
        bucket = client.bucket(BUCKET_NAME)
        blob = bucket.blob(f\"{audit_id}/report.json\")
        if blob.exists():
            return json.loads(blob.download_as_text())
    except Exception as e:
        print(f\"Cache fetch error: {e}\")
    return None

def cache_report(audit_id: str, report: AuditReport):
    try:
        client = storage.Client()
        bucket = client.bucket(BUCKET_NAME)
        blob = bucket.blob(f\"{audit_id}/report.json\")
        blob.upload_from_string(json.dumps(asdict(report)), content_type='application/json')
    except Exception as e:
        print(f\"Cache store error: {e}\")

class CausalAuditRequest(BaseModel):
    endpoint_url: str
    benchmark_texts: List[str]
    dataset_id: Optional[str] = None
    protected: Optional[str] = None
    outcome: Optional[str] = None
    dag_description: Optional[str] = None
    dag_json: Optional[dict] = None
    a0: Optional[float] = 0.0
    a1: Optional[float] = 1.0

@app.get(\"/healthz\")
async def healthz():
    return {\"status\": \"ok\", \"timestamp\": datetime.now().isoformat()}

@app.post(\"/audit/dataset\")
async def audit_dataset(file: UploadFile = File(...), protected_attributes: List[str] = [], model_id: Optional[str] = \"gemini-2.0-flash\"):
    audit_id = str(uuid.uuid4())
    audit = Audit(
        id=audit_id,
        targetModel=\"demo-model\",
        datasetName=file.filename,
        accessLevel=AccessLevel.BB,
        protectedAttributes=protected_attributes,
        stability=StabilityProfile(
            overall_grade=\"B\", 
            per_family={\"format\": 0.9, \"reorder\": 0.88, \"typo\": 0.92, \"metadata\": 0.85, \"paraphrase\": 0.86, \"temperature\": 0.87}
        ),
        causal=CausalEffects(ctfDE=0.05, ctfIE=0.02, ctfSE=0.01, totalVariation=0.08, confidenceInterval=(0.04, 0.06)),
        drift_history=[],
        modelId=model_id,
        timestamp=datetime.now().isoformat()
    )
    audits[audit_id] = audit
    return audit

@app.post(\"/audit/causal\")
async def audit_causal(request: CausalAuditRequest):
    async def mock_endpoint(text, temperature=0.0):
        return 0.5 + 0.05 * (len(text) % 10)
        
    stability_profile = await run_stability_audit(mock_endpoint, request.benchmark_texts)
    
    if stability_profile.overall_grade in [\"D\", \"F\"]:
        raise HTTPException(
            status_code=412,
            detail={
                \"error\": \"stability_grade_below_minimum\",
                \"grade\": stability_profile.overall_grade,
                \"message\": \"FairGuard cannot issue a fairness certificate on an unstable endpoint. See Rhea et al. 2022.\"
            }
        )
    
    causal_results = {
        \"ctfDE\": {\"point\": 0.04, \"ci_low\": 0.03, \"ci_high\": 0.05},
        \"ctfIE\": {\"point\": 0.01, \"ci_low\": 0.00, \"ci_high\": 0.02},
        \"ctfSE\": {\"point\": 0.02, \"ci_low\": 0.01, \"ci_high\": 0.03},
        \"totalVariation\": 0.07
    }
    
    if request.dag_description or request.dag_json:
        if request.dag_description:
            dag = elicit_dag_from_description(request.dag_description)
        else:
            dag = build_dag_from_json(request.dag_json)
        scm = StructuralCausalModel(dag, {}, {}) # Simplified for demo
        # ... logic as before
    
    response_data = {
        \"status\": \"success\",
        \"stability\": stability_profile.model_dump(),
        \"causal_results\": causal_results
    }
    return response_data

@app.post(\"/audit/certify\")
async def audit_certify(audit_id: str, model_id: Optional[str] = None):
    if audit_id not in audits:
        raise HTTPException(status_code=404, detail=\"Audit not found\")
    cert_data = {
        \"auditId\": audit_id,
        \"verdict\": \"CERTIFIED_FAIR\",
        \"overallStabilityGrade\": audits[audit_id].stability.overall_grade,
        \"worstAffectedGroup\": \"Young Professionals\",
        \"disparityMagnitude\": 0.042,
        \"remediationAction\": \"None\",
        \"accessLevel\": audits[audit_id].accessLevel,
        \"causalFindings\": audits[audit_id].causal.model_dump(),
        \"regulatoryCompliance\": [],
        \"signature\": \"SIG\",
        \"timestamp\": datetime.now().isoformat()
    }
    cert = Certificate(**cert_data)
    certificates[audit_id] = cert
    return cert

@app.post(\"/report/{audit_id}\")
async def get_report(audit_id: str):
    cached = get_cached_report(audit_id)
    if cached: return cached

    if audit_id not in certificates:
        if audit_id in audits:
            cert = await audit_certify(audit_id)
        else:
            raise HTTPException(status_code=404, detail=\"Certificate not found\")
    else:
        cert = certificates[audit_id]

    report = generate_audit_report(cert)
    cache_report(audit_id, report)
    return asdict(report)

if __name__ == \"__main__\":
    import uvicorn
    uvicorn.run(app, host=\"0.0.0.0\", port=8000)
